@@ -1,6 +1,6 @@
 {
   "name": "simsage-admin",
-  "version": "5.6.2",
+  "version": "7.0.5",
   "private": true,
   "description": "SimSage: search less; find more",
   "repository": "git@gitlab.com:peter3125/simsage.git",
@@ -15,38 +15,41 @@
     "eject": "react-scripts eject"
   },
   "dependencies": {
-    "@date-io/core": "^2.8.0",
-    "@date-io/date-fns": "^2.8.0",
-    "@material-ui/core": "^4.11.0",
-    "@material-ui/icons": "^4.9.1",
+    "@date-io/core": "^2.11.0",
+    "@date-io/date-fns": "^2.10.11",
     "axios": "^0.21.1",
-    "chart.js": "^2.8.0",
-    "cleave.js": "^1.5.3",
-    "core-js": "^3.6.5",
-    "date-fns": "^2.15.0",
-    "downshift": "^6.0.2",
-    "luxon": "^1.24.1",
-    "moment": "^2.27.0",
-    "npm": "^7.3.0",
-    "query-string": "^6.13.1",
-    "react": "^16.13.1",
-    "react-app-polyfill": "^1.0.6",
-    "react-chartjs-2": "^2.9.0",
+    "bootstrap": "^5.0.2",
+    "cleave.js": "^1.6.0",
+    "core-js": "^3.15.2",
+    "date-fns": "^2.22.1",
+    "downshift": "^6.1.3",
+    "history": "^5.0.0",
+    "jquery": "^3.6.0",
+    "luxon": "^2.0.1",
+    "moment": "^2.29.1",
+    "npm": "^7.19.1",
+    "@popperjs/core": "^2.9.2",
+    "query-string": "^7.0.0",
+    "react": "^17.0.2",
+    "react-app-polyfill": "^2.0.0",
     "react-datepicker": "^3.1.3",
-    "react-dom": "^16.13.1",
-    "react-redux": "^7.2.0",
+    "react-chartjs-2": "^2.9.0",
+    "react-dom": "^17.0.2",
+    "react-redux": "^7.2.4",
+    "react-rnd": "^10.3.3",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
     "react-router-redux": "^4.0.8",
-    "react-scripts": "^4.0.1",
+    "react-scripts": "^4.0.3",
+    "react-scroll": "^1.8.2",
     "react-stomp": "^5.0.0",
-    "redux": "^4.0.5",
+    "redux": "^4.1.0",
     "redux-logger": "^3.0.6",
     "redux-persist": "^6.0.0",
     "redux-thunk": "^2.3.0",
-    "styled-components": "^5.2.1",
-    "typeface-roboto": "0.0.75",
-    "url-loader": "^4.1.0"
+    "styled-components": "^5.3.0",
+    "typeface-roboto": "^1.1.13",
+    "url-loader": "^4.1.1"
   },
   "eslintConfig": {
     "extends": "react-app"
