@@ -1,7 +1,7 @@
 
 window.ENV = {
     // SimSage platform version (used for display only in UI)
-    version: '5.6.2',
+    version: '7.0.5',
     // api version of api_base
     api_version: 1,
     // is this a production build or not?
