@@ -1,92 +1,16 @@
 import React, {Component} from 'react';
 
-import Table from '@material-ui/core/Table';
-import TableBody from '@material-ui/core/TableBody';
-import TableCell from '@material-ui/core/TableCell';
-import TableHead from '@material-ui/core/TableHead';
-import TableRow from '@material-ui/core/TableRow';
-import Paper from '@material-ui/core/Paper';
-import TablePagination from '@material-ui/core/TablePagination';
-
 import {DomainDialog} from './domain_dialog'
 
 import {connect} from "react-redux";
 import {bindActionCreators} from "redux";
 import {appCreators} from "../actions/appActions";
 
+import '../css/domains.css';
+import {Pagination} from "../common/pagination";
+
 const empty_domain= {domainId: '', domainName: '', userName: '', password: '', serverIp: '', basePath: '', portNumber: 389, sslOn: false, schedule: '' };
 
-const styles = {
-    pageWidth: {
-        width: '900px',
-    },
-    tableLight: {
-    },
-    tableDark: {
-        background: '#d0d0d0',
-    },
-    tab: {
-        backgroundColor: '#f8f8f8',
-        color: '#000',
-    },
-    addImage: {
-        width: '25px',
-        cursor: 'pointer',
-        color: "green",
-    },
-    knowledgeSelect: {
-        padding: '5px',
-        marginBottom: '40px',
-    },
-    lhs: {
-        float: 'left',
-        width: '150px',
-        marginTop: '-10px',
-        color: '#aaa',
-    },
-    rhs: {
-        float: 'left',
-    },
-    tableHeaderStyle: {
-        background: '#555',
-        fontSize: '0.95em',
-        color: '#fff',
-    },
-    label: {
-        marginTop: '20px',
-        color: '#555',
-    },
-    linkButton: {
-        float: 'left',
-        padding: '10px',
-        color: '#888',
-        cursor: 'pointer',
-    },
-    imageButton: {
-        float: 'right',
-        padding: '10px',
-        color: '#888',
-        cursor: 'pointer',
-    },
-    download: {
-        cursor: 'pointer',
-        fontSize: '0.9em',
-        fontWeight: '600',
-        height: '24px',
-    },
-    downloadType: {
-        fontSize: '0.9em',
-        height: '24px',
-    },
-    downloadImage: {
-        width: '24px',
-    },
-    labelHeader: {
-        marginTop: '20px',
-        marginBottom: '5px',
-        fontSize: '0.8em',
-    },
-};
 
 export class Domains extends Component {
     constructor(props) {
@@ -97,7 +21,7 @@ export class Domains extends Component {
             page: 0,
             // dialog
             open: false,
-            title: '',
+            title: 'Create new domain',
             selected_domain: {},
             domain_ask: null,
         };
@@ -124,14 +48,14 @@ export class Domains extends Component {
         return paginated_list;
     }
     addNewDomain() {
-        this.setState({open: true, selected_domain: empty_domain, title: 'Create New Domain'});
+        this.setState({open: true, selected_domain: empty_domain, title: 'Create new domain'});
     }
     onUpdate(domain) {
         this.setState({selected_domain: domain});
     }
     editDomain(domain) {
         if (domain) {
-            this.setState({open: true, selected_domain: { ...empty_domain, ...domain}, title: 'Edit Domain'});
+            this.setState({open: true, selected_domain: { ...empty_domain, ...domain}, title: 'Edit domain'});
         }
     }
     deleteDomainAsk(domain) {
@@ -191,59 +115,59 @@ export class Domains extends Component {
                 {
                     this.isVisible() &&
 
-                    <Paper style={styles.pageWidth}>
-                        <Table>
-                            <TableHead>
-                                <TableRow className='table-header'>
-                                    <TableCell className='table-header'>name</TableCell>
-                                    <TableCell className='table-header'>path</TableCell>
-                                    <TableCell className='table-header'>objects</TableCell>
-                                    <TableCell className='table-header'>actions</TableCell>
-                                </TableRow>
-                            </TableHead>
-                            <TableBody style={theme === 'light' ? styles.tableLight : styles.tableDark}>
+                    <div className="page-width">
+                        <table className="table">
+                            <thead>
+                                <tr className='table-header'>
+                                    <th scope="col" className='table-header'>name</th>
+                                    <th scope="col" className='table-header'>path</th>
+                                    <th scope="col" className='table-header'>objects</th>
+                                    <th scope="col" className='table-header'>actions</th>
+                                </tr>
+                            </thead>
+                            <tbody className={theme === 'light' ? "table-light" : "table-dark"}>
                                 {
                                     this.getDomains().map((domain) => {
                                         return (
-                                            <TableRow key={domain.domainId}>
-                                                <TableCell>
-                                                    <div style={styles.label}>{domain.domainName}</div>
-                                                </TableCell>
-                                                <TableCell>
-                                                    <div style={styles.label}>{domain.basePath}</div>
-                                                </TableCell>
-                                                <TableCell>
-                                                    <div style={styles.label}>{domain.numObjects}</div>
-                                                </TableCell>
-                                                <TableCell>
-                                                    <div style={styles.linkButton} onClick={() => this.editDomain(domain)}>
-                                                        <img src="../images/edit.svg" style={styles.downloadImage} title="edit domain" alt="edit"/>
+                                            <tr key={domain.domainId}>
+                                                <td>
+                                                    <div className="label">{domain.domainName}</div>
+                                                </td>
+                                                <td>
+                                                    <div className="label">{domain.basePath}</div>
+                                                </td>
+                                                <td>
+                                                    <div className="label">{domain.numObjects}</div>
+                                                </td>
+                                                <td>
+                                                    <div className="label-button" onClick={() => this.editDomain(domain)}>
+                                                        <img src="../images/edit.svg" className="download-image" title="edit domain" alt="edit"/>
                                                     </div>
-                                                    <div style={styles.linkButton} onClick={() => this.deleteDomainAsk(domain)}>
-                                                        <img src="../images/delete.svg" style={styles.downloadImage} title="remove domain" alt="remove"/>
+                                                    <div className="label-button" onClick={() => this.deleteDomainAsk(domain)}>
+                                                        <img src="../images/delete.svg" className="download-image" title="remove domain" alt="remove"/>
                                                     </div>
-                                                </TableCell>
-                                            </TableRow>
+                                                </td>
+                                            </tr>
                                         )
                                     })
                                 }
-                                <TableRow>
-                                    <TableCell/>
-                                    <TableCell/>
-                                    <TableCell/>
-                                    <TableCell>
+                                <tr>
+                                    <td/>
+                                    <td/>
+                                    <td/>
+                                    <td>
                                         {this.props.selected_organisation_id.length > 0 && this.props.selected_knowledgebase_id.length > 0 &&
-                                        <div style={styles.imageButton} onClick={() => this.addNewDomain()}><img
-                                            style={styles.addImage} src="../images/add.svg" title="add new domain"
+                                        <div className="image-button" onClick={() => this.addNewDomain()}><img
+                                            className="add-image" src="../images/add.svg" title="add new domain"
                                             alt="add new domain"/></div>
                                         }
-                                    </TableCell>
-                                </TableRow>
-                            </TableBody>
-                        </Table>
+                                    </td>
+                                </tr>
+                            </tbody>
+                        </table>
 
-                        <TablePagination
-                            style={theme === 'light' ? styles.tableLight : styles.tableDark}
+                        <Pagination
+                            theme={theme}
                             rowsPerPageOptions={[5, 10, 25]}
                             component="div"
                             count={this.props.domain_list.length}
@@ -255,11 +179,11 @@ export class Domains extends Component {
                             nextIconButtonProps={{
                                 'aria-label': 'Next Page',
                             }}
-                            onChangePage={(event, page) => this.changePage(page)}
-                            onChangeRowsPerPage={(event) => this.changePageSize(event.target.value)}
+                            onChangePage={(page) => this.changePage(page)}
+                            onChangeRowsPerPage={(rows) => this.changePageSize(rows)}
                         />
 
-                    </Paper>
+                    </div>
                 }
 
             </div>
