@@ -1,72 +1,10 @@
 import React, {Component} from 'react';
-import Grid from '@material-ui/core/Grid';
 
-import {Switch} from "@material-ui/core";
 import {connect} from "react-redux";
 import {bindActionCreators} from "redux";
 import {appCreators} from "../actions/appActions";
 
-const styles = {
-    background: {
-        margin: '5px',
-        backgroundColor: '#fefefe',
-        padding: '15px',
-        height: '55px',
-        fontFamily: 'Tahoma',
-        fontWeight: '500',
-        color: '#888',
-        borderRadius: '4px',
-    },
-    logo_box: {
-        position: 'absolute',
-        left: '10px',
-        top: '10px'
-    },
-    logo: {
-        float: 'left',
-        width: '200px',
-    },
-    projectTitle: {
-        fontSize: '1.1em',
-        float: 'left',
-        marginLeft: '50px',
-        marginTop: '25px',
-    },
-    homeImageContainer: {
-        float: 'left',
-    },
-    themeSelect: {
-        marginRight: '20px',
-        marginTop: '4px',
-        float: 'left',
-    },
-    themeSelectLoggedIn: {
-        marginTop: '-10px',
-        marginRight: '10px',
-        float: 'left',
-    },
-    homeImage: {
-        width: '32px',
-    },
-    signOutImageContainer: {
-        paddingTop: '5px',
-    },
-    signOutImage: {
-        width: '22px',
-    },
-    bookButton: {
-        float: 'right',
-        marginRight: '20px',
-    },
-    bookImage: {
-        width: '32px',
-    },
-    versionText: {
-        fontSize: '0.8em',
-        marginLeft: '-20px',
-    }
-};
-
+import '../css/app-menu.css';
 
 export class AppMenu extends Component {
     constructor(props){
@@ -111,46 +49,40 @@ export class AppMenu extends Component {
     render() {
         const theme = this.props.theme;
         return (
-            <div className={theme + " menu-padding no-select"}>
-                <div style={styles.logo_box}>
-                    <img alt="SimSage" title="Search less; find more." style={styles.logo} src={this.getLogo()} onClick={() => this.goWeb()} />
+            <div className={"app-menu menu-padding no-select"}>
+                <div className="logo-box">
+                    <img alt="SimSage" title="Search less; find more." className="logo" src={this.getLogo()} onClick={() => this.goWeb()} />
                 </div>
-                <Grid container spacing={1}>
-                    <Grid item xs={2} />
-                    <Grid item xs={4}>
-                        <span style={styles.projectTitle}>{this.state.title}</span>
-                    </Grid>
-                    <Grid item xs={4} />
-                    <Grid item xs={2}>
-                        {
-                            <div style={this.state.signed_in ? styles.themeSelectLoggedIn : styles.themeSelect}>
-                                <Switch
-                                    checked={this.props.theme === 'light'}
-                                    onChange={() => this.flipTheme()}
-                                    title="change the SimSage theme from dark to light or vice versa"
-                                    color="primary"
-                                    name="checkedTheme"
-                                    inputProps={{ 'aria-label': 'primary checkbox' }}
-                                />
+                <div>
+
+                    {
+                        !this.state.signed_in &&
+                        <span className="home-image-container">
+                            <img src={theme === 'light' ? "../images/home.svg" : "../images/home-light.svg"} alt="home" title="home" onClick={() => this.goWeb()}
+                                 className="home-image" />
+                             <div className="version-text">version {window.ENV.version}</div>
+                        </span>
+                    }
+                    {
+                        this.state.signed_in &&
+                        <div className="sign-out-image-container">
+                            <img src={theme === 'light' ? "../images/sign-out.svg" : "../images/sign-out-light.svg"} alt="sign-out" title="sign-out"
+                                 onClick={() => { this.signOut() }} className="sign-out-image" />
+                        </div>
+                    }
+
+                    {
+                        <div className={this.state.signed_in ? "theme-select-logged-in" : "theme-select"}>
+                            <div className="form-check form-switch">
+                                <input className="form-check-input" type="checkbox" id="flexSwitchCheckChecked" checked={this.props.theme === 'light'}
+                                       title="change the SimSage theme from dark to light or vice versa"
+                                       onChange={() => this.flipTheme()} />
                             </div>
-                        }
-                        {
-                            !this.state.signed_in &&
-                            <span style={styles.homeImageContainer}>
-                                <img src={theme === 'light' ? "../images/home.svg" : "../images/home-light.svg"} alt="home" title="home" onClick={() => this.goWeb()}
-                                     style={styles.homeImage}/>
-                                 <div style={styles.versionText}>version {window.ENV.version}</div>
-                            </span>
-                        }
-                        {
-                            this.state.signed_in &&
-                            <span style={styles.signOutImageContainer}>
-                                <img src={theme === 'light' ? "../images/sign-out.svg" : "../images/sign-out-light.svg"} alt="sign-out" title="sign-out"
-                                     onClick={() => { this.signOut() }} style={styles.signOutImage}/>
-                            </span>
-                        }
-                    </Grid>
-                </Grid>
+                        </div>
+                    }
+
+                </div>
+                <br clear="both" />
             </div>
         );
     }
