@@ -1,13 +1,5 @@
-
 import React, {Component} from 'react';
 
-import Dialog from '@material-ui/core/Dialog';
-import DialogActions from '@material-ui/core/DialogActions';
-import DialogContent from '@material-ui/core/DialogContent';
-import DialogTitle from '@material-ui/core/DialogTitle';
-import Button from '@material-ui/core/Button';
-import Grid from "@material-ui/core/Grid";
-import TextField from "@material-ui/core/TextField";
 
 export class OperatorTeach extends Component {
     constructor(props) {
@@ -38,64 +30,47 @@ export class OperatorTeach extends Component {
         if (this.state.has_error) {
             return <h1>operator-teach.js: Something went wrong.</h1>;
         }
-        const theme = this.props.theme;
+        if (!this.props.open)
+            return (<div />);
         return (
-            <Dialog aria-labelledby="alert-dialog-title"
-                    aria-describedby="alert-dialog-description"
-                    disableBackdropClick={true}
-                    disableEscapeKeyDown={true}
-                    open={this.props.open}
-                    fullWidth={true}
-                    maxWidth="lg"
-                    onClose={() => this.handleCancel()} >
+            <div className="modal" tabIndex="-1" role="dialog" style={{display: "inline"}}>
+                <div className={"modal-dialog modal-dialog-centered modal-xl"} role="document">
+                    <div className="modal-content shadow p-3 mb-5 bg-white rounded">
 
-                <DialogTitle id="alert-dialog-title" className={theme}>Teach SimSage?</DialogTitle>
-                <DialogContent className={theme}>
-                    <Grid container spacing={1}>
+                        <div className="modal-header">Teach SimSage?</div>
+                        <div className="modal-body">
+                            <div>
 
-                        <Grid item xs={1} />
-                        <Grid item xs={1}>
-                            <div>question</div>
-                        </Grid>
-                        <Grid item xs={9}>
-                            {this.props.question}
-                        </Grid>
-                        <Grid item xs={1} />
+                                <div>question</div>
+                                <div>
+                                    {this.props.question}
+                                </div>
 
+                                <div>answer</div>
+                                <div>
+                                    {this.props.answer}
+                                </div>
 
-                        <Grid item xs={1} />
-                        <Grid item xs={1}>
-                            <div>answer</div>
-                        </Grid>
-                        <Grid item xs={9}>
-                            {this.props.answer}
-                        </Grid>
-                        <Grid item xs={1} />
+                                <div>links</div>
+                                <div>
+                                    <textarea
+                                        onChange={(event) => this.setState({links: event.target.value})}
+                                        placeholder="links (images and page-links, separated by commas or spaces)"
+                                        rows={5}
+                                        value={this.state.links}
+                                    />
+                                </div>
 
-                        <Grid item xs={1} />
-                        <Grid item xs={1}>
-                            <div>links</div>
-                        </Grid>
-                        <Grid item xs={9}>
-                            <TextField
-                                onChange={(event) => this.setState({links: event.target.value})}
-                                placeholder="links (images and page-links, separated by commas or spaces)"
-                                multiline={true}
-                                rows={5}
-                                variant="outlined"
-                                fullWidth={true}
-                                value={this.state.links}
-                            />
-                        </Grid>
-                        <Grid item xs={1} />
+                            </div>
+                        </div>
+                        <div className="modal-footer">
+                            <button className="btn btn-primary btn-block"  onClick={() => this.handleCancel()}>Cancel</button>
+                            <button className="btn btn-primary btn-block"  onClick={() => this.handleSave()}>Save</button>
+                        </div>
 
-                    </Grid>
-                </DialogContent>
-                <DialogActions className={theme}>
-                    <Button color="primary" onClick={() => this.handleCancel()}>Cancel</Button>
-                    <Button variant="contained" color="secondary" onClick={() => this.handleSave()}>Save</Button>
-                </DialogActions>
-            </Dialog>
+                    </div>
+                </div>
+            </div>
         );
     }
 }
