@@ -1,133 +1,11 @@
 import React from 'react';
 
-import Table from '@material-ui/core/Table';
-import TableBody from '@material-ui/core/TableBody';
-import TableCell from '@material-ui/core/TableCell';
-import TableHead from '@material-ui/core/TableHead';
-import TableRow from '@material-ui/core/TableRow';
-import Paper from '@material-ui/core/Paper';
-
 import {SemanticEdit} from "./semantic-edit";
-import TablePagination from "@material-ui/core/TablePagination";
 
 import {connect} from "react-redux";
 import {bindActionCreators} from "redux";
 import {appCreators} from "../actions/appActions";
-
-
-
-const styles = {
-    tableStyle: {
-        minWidth: '900px',
-        width: '900px',
-    },
-    tableLight: {
-    },
-    tableDark: {
-        background: '#d0d0d0',
-    },
-    smallTableHeaderStyle: {
-        background: '#555',
-        fontSize: '0.95em',
-        color: '#fff',
-        minWidth: '20px',
-    },
-    actionTableHeaderStyle: {
-        background: '#555',
-        fontSize: '0.95em',
-        color: '#fff',
-        minWidth: '100px',
-    },
-    tableHeaderStyle: {
-        background: '#555',
-        fontSize: '0.95em',
-        color: '#fff',
-        minWidth: '400px',
-        width: '600px',
-    },
-    searchTableHeaderStyle: {
-        background: '#555',
-        fontSize: '0.95em',
-        color: '#fff',
-        minWidth: '200px',
-        width: '340px',
-    },
-    knowledgeSelect: {
-        padding: '5px',
-        marginBottom: '40px',
-    },
-    lhs: {
-        float: 'left',
-        width: '150px',
-        marginTop: '-10px',
-        color: '#aaa',
-    },
-    rhs: {
-        float: 'left',
-    },
-    label: {
-        padding: '10px',
-        color: '#555',
-    },
-    gridWidth: {
-        width: '900px',
-    },
-    hr: {
-        border: '0.1px solid #f0f0f0',
-        width: '100%',
-    },
-    imageButton: {
-        float: 'right',
-        padding: '10px',
-        color: '#888',
-        cursor: 'pointer',
-    },
-    addImage: {
-        width: '25px',
-    },
-    linkButton: {
-        float: 'left',
-        padding: '10px',
-        color: '#888',
-        cursor: 'pointer',
-    },
-    dlImageSize: {
-        width: '24px',
-    },
-    search: {
-        marginTop: '2px',
-        marginLeft: '15px',
-        width: '18px',
-        color: '#000',
-    },
-    floatLeftLabel: {
-        float: 'left',
-        marginRight: '6px',
-        marginTop: '4px',
-        fontSize: '0.9em',
-        fontWeight: '500',
-    },
-    searchFloatLeft: {
-        float: 'left',
-    },
-    text: {
-        border: '1px solid #808080',
-        borderRadius: '4px',
-        padding: '4px',
-        width: '280px',
-    },
-    floatLeft: {
-        float: 'left',
-    },
-    manualUploadSection: {
-        marginTop: '50px',
-    },
-    findBox: {
-        padding: '10px',
-        marginBottom: '5px',
-        float: 'right',
-    },
-};
+import {Pagination} from "../common/pagination";
 
 
 export class Semantics extends React.Component {
@@ -246,21 +124,20 @@ export class Semantics extends React.Component {
 
                 {
                     this.isVisible() &&
-
-                    <div style={styles.findBox}>
-                        <div style={styles.floatLeftLabel}>find semantics</div>
-                        <div style={styles.searchFloatLeft}>
-                            <input type="text" value={this.props.semantic_filter} autoFocus={true} style={styles.text} className={theme}
+                    <div className="filter-find-box">
+                        <span className="filter-label">find semantics</span>
+                        <span className="filter-find-text">
+                            <input type="text" value={this.props.semantic_filter} autoFocus={true} className={theme}
                                    onKeyPress={(event) => this.handleSearchTextKeydown(event)}
                                    onChange={(event) => {
                                        this.props.setSemanticFilter(event.target.value);
                                    }}/>
-                        </div>
-                        <div style={styles.floatLeft}>
-                            <img style={styles.search}
+                        </span>
+                        <span className="filter-find-image">
+                            <img className="image-size"
                                  onClick={() => this.props.getSemantics()}
                                  src="../images/dark-magnifying-glass.svg" title="search" alt="search"/>
-                        </div>
+                        </span>
                     </div>
                 }
 
@@ -268,70 +145,63 @@ export class Semantics extends React.Component {
 
                 {
                     this.isVisible() &&
-                    <Paper>
-                        <Table style={styles.tableStyle}>
-                            <TableHead>
-                                <TableRow>
-                                    <TableCell className='table-header'>word</TableCell>
-                                    <TableCell className='table-header'>semantic</TableCell>
-                                    <TableCell className='table-header'>actions</TableCell>
-                                </TableRow>
-                            </TableHead>
-                            <TableBody style={theme === 'light' ? styles.tableLight : styles.tableDark}>
+                    <div>
+                        <table className="table">
+                            <thead>
+                                <tr className='table-header'>
+                                    <th className='table-header'>word</th>
+                                    <th className='table-header'>semantic</th>
+                                    <th className='table-header'>actions</th>
+                                </tr>
+                            </thead>
+                            <tbody>
                                 {
                                     this.getSemanticList().map((semantic) => {
                                         return (
-                                            <TableRow key={semantic.word + ":" + semantic.semantic}>
-                                                <TableCell>
-                                                    <div style={styles.label}>{semantic.word}</div>
-                                                </TableCell>
-                                                <TableCell>
-                                                    <div style={styles.label}>{semantic.semantic}</div>
-                                                </TableCell>
-                                                <TableCell>
-                                                    <div style={styles.linkButton} onClick={() => this.editSemantic(semantic)}>
-                                                        <img src="../images/edit.svg" style={styles.dlImageSize} title="edit semantic" alt="edit"/>
+                                            <tr key={semantic.word + ":" + semantic.semantic}>
+                                                <td>
+                                                    <div>{semantic.word}</div>
+                                                </td>
+                                                <td>
+                                                    <div>{semantic.semantic}</div>
+                                                </td>
+                                                <td>
+                                                    <div className="link-button" onClick={() => this.editSemantic(semantic)}>
+                                                        <img src="../images/edit.svg" className="image-size" title="edit semantic" alt="edit"/>
                                                     </div>
-                                                    <div style={styles.linkButton} onClick={() => this.deleteSemanticAsk(semantic)}>
-                                                        <img src="../images/delete.svg" style={styles.dlImageSize} title="remove semantic" alt="remove"/>
+                                                    <div className="link-button" onClick={() => this.deleteSemanticAsk(semantic)}>
+                                                        <img src="../images/delete.svg" className="image-size" title="remove semantic" alt="remove"/>
                                                     </div>
-                                                </TableCell>
-                                            </TableRow>
+                                                </td>
+                                            </tr>
                                         )
                                     })
                                 }
-                                <TableRow>
-                                    <TableCell/>
-                                    <TableCell/>
-                                    <TableCell>
+                                <tr>
+                                    <td/>
+                                    <td/>
+                                    <td>
                                         {this.isVisible() &&
-                                        <div style={styles.imageButton} onClick={() => this.newSemantic()}><img
-                                            style={styles.addImage} src="../images/add.svg" title="new semantic"
-                                            alt="new semantic"/></div>
+                                            <div className="link-button" onClick={() => this.newSemantic()}><img
+                                                className="image-size" src="../images/add.svg" title="new semantic" alt="new semantic"/></div>
                                         }
-                                    </TableCell>
-                                </TableRow>
-                            </TableBody>
-                        </Table>
+                                    </td>
+                                </tr>
+                            </tbody>
+                        </table>
 
-                        <TablePagination
-                            style={theme === 'light' ? styles.tableLight : styles.tableDark}
+                        <Pagination
+                            theme={theme}
                             rowsPerPageOptions={[5, 10, 25]}
                             component="div"
                             count={this.props.num_semantics}
                             rowsPerPage={this.props.semantic_page_size}
                             page={this.props.semantic_page}
-                            backIconButtonProps={{
-                                'aria-label': 'Previous Page',
-                            }}
-                            nextIconButtonProps={{
-                                'aria-label': 'Next Page',
-                            }}
-                            onChangePage={(event, page) => this.props.setSemanticPage(page)}
-                            onChangeRowsPerPage={(event) => this.props.setSemanticPageSize(event.target.value)}
+                            onChangePage={(page) => this.props.setSemanticPage(page)}
+                            onChangeRowsPerPage={(rows) => this.props.setSemanticPageSize(rows)}
                         />
 
-                    </Paper>
+                    </div>
                 }
 
             </div>
