@@ -1,132 +1,13 @@
 import React from 'react';
 
-import Table from '@material-ui/core/Table';
-import TableBody from '@material-ui/core/TableBody';
-import TableCell from '@material-ui/core/TableCell';
-import TableHead from '@material-ui/core/TableHead';
-import TableRow from '@material-ui/core/TableRow';
-import Paper from '@material-ui/core/Paper';
-
 import {SynSetEdit} from "./synset-edit";
-import TablePagination from "@material-ui/core/TablePagination";
 
 import {connect} from "react-redux";
 import {bindActionCreators} from "redux";
 import {appCreators} from "../actions/appActions";
+import {Pagination} from "../common/pagination";
 
-
-const styles = {
-    tableStyle: {
-        minWidth: '800px',
-        width: '900px',
-    },
-    tableLight: {
-    },
-    tableDark: {
-        background: '#d0d0d0',
-    },
-    smallTableHeaderStyle: {
-        background: '#555',
-        fontSize: '0.95em',
-        color: '#fff',
-        minWidth: '20px',
-    },
-    actionTableHeaderStyle: {
-        background: '#555',
-        fontSize: '0.95em',
-        color: '#fff',
-        minWidth: '100px',
-    },
-    tableHeaderStyle: {
-        background: '#555',
-        fontSize: '0.95em',
-        color: '#fff',
-        minWidth: '400px',
-        width: '600px',
-    },
-    searchTableHeaderStyle: {
-        background: '#555',
-        fontSize: '0.95em',
-        color: '#fff',
-        minWidth: '200px',
-        width: '340px',
-    },
-    knowledgeSelect: {
-        padding: '5px',
-        marginBottom: '40px',
-    },
-    lhs: {
-        float: 'left',
-        width: '150px',
-        marginTop: '-10px',
-        color: '#aaa',
-    },
-    rhs: {
-        float: 'left',
-    },
-    label: {
-        padding: '10px',
-        color: '#555',
-    },
-    gridWidth: {
-        width: '900px',
-    },
-    hr: {
-        border: '0.1px solid #f0f0f0',
-        width: '100%',
-    },
-    imageButton: {
-        float: 'right',
-        padding: '10px',
-        color: '#888',
-        cursor: 'pointer',
-    },
-    addImage: {
-        width: '25px',
-    },
-    linkButton: {
-        float: 'left',
-        padding: '10px',
-        color: '#888',
-        cursor: 'pointer',
-    },
-    dlImageSize: {
-        width: '24px',
-    },
-    search: {
-        marginTop: '2px',
-        marginLeft: '15px',
-        width: '18px',
-        color: '#000',
-    },
-    floatLeftLabel: {
-        float: 'left',
-        marginRight: '6px',
-        marginTop: '4px',
-        fontSize: '0.9em',
-        fontWeight: '500',
-    },
-    searchFloatLeft: {
-        float: 'left',
-    },
-    text: {
-        border: '1px solid #808080',
-        borderRadius: '4px',
-        padding: '4px',
-        width: '280px',
-    },
-    floatLeft: {
-        float: 'left',
-    },
-    manualUploadSection: {
-        marginTop: '50px',
-    },
-    findBox: {
-        padding: '10px',
-        marginBottom: '5px',
-        float: 'right',
-    },
-};
+import '../css/synset.css';
 
 
 export class SynSets extends React.Component {
@@ -212,7 +93,7 @@ export class SynSets extends React.Component {
     render() {
         const theme = this.props.theme;
         return (
-            <div>
+            <div className="synset-page">
                 <SynSetEdit open={this.state.synSet_edit}
                             theme={theme}
                             synSet={this.state.synSet}
@@ -222,20 +103,20 @@ export class SynSets extends React.Component {
                 {
                     this.isVisible() &&
 
-                    <div style={styles.findBox}>
-                        <div style={styles.floatLeftLabel}>filter syn-sets</div>
-                        <div style={styles.searchFloatLeft}>
-                            <input type="text" value={this.props.synset_filter} autoFocus={true} style={styles.text} className={theme}
+                    <div className="filter-find-box">
+                        <span className="filter-label">find synsets</span>
+                        <span className="filter-find-text">
+                            <input type="text" value={this.props.synset_filter} autoFocus={true}
                                    onKeyPress={(event) => this.handleSynSetFilterKeydown(event)}
                                    onChange={(event) => {
                                        this.props.setSynSetFilter(event.target.value)
                                    }}/>
-                        </div>
-                        <div style={styles.floatLeft}>
-                            <img style={styles.search}
+                        </span>
+                        <span className="filter-find-image">
+                            <img className="image-size"
                                  onClick={() => this.props.findSynSets()}
                                  src="../images/dark-magnifying-glass.svg" title="search" alt="search"/>
-                        </div>
+                        </span>
                     </div>
                 }
 
@@ -243,67 +124,61 @@ export class SynSets extends React.Component {
 
                 {
                     this.isVisible() &&
-                    <Paper>
-                        <Table style={styles.tableStyle}>
-                            <TableHead>
-                                <TableRow>
-                                    <TableCell className='table-header'>syn-set</TableCell>
-                                    <TableCell className='table-header'>actions</TableCell>
-                                </TableRow>
-                            </TableHead>
-                            <TableBody style={theme === 'light' ? styles.tableLight : styles.tableDark}>
+                    <div>
+                        <table className="table">
+                            <thead>
+                                <tr className='table-header'>
+                                    <th className='table-header'>syn-set</th>
+                                    <th className='table-header'>actions</th>
+                                </tr>
+                            </thead>
+                            <tbody>
                                 {
                                     this.props.synset_list.map((synSet) => {
                                         return (
-                                            <TableRow key={synSet.word}>
-                                                <TableCell>
-                                                    <div style={styles.label}>{synSet.word}</div>
-                                                </TableCell>
-                                                <TableCell>
-                                                    <div style={styles.linkButton} onClick={() => this.editSynSet(synSet)}>
-                                                        <img src="../images/edit.svg" style={styles.dlImageSize} title="edit syn-set" alt="edit"/>
+                                            <tr key={synSet.word}>
+                                                <td>
+                                                    <div className="synset-label">{synSet.word}</div>
+                                                </td>
+                                                <td>
+                                                    <div className="link-button" onClick={() => this.editSynSet(synSet)}>
+                                                        <img src="../images/edit.svg" className="image-size" title="edit syn-set" alt="edit"/>
                                                     </div>
-                                                    <div style={styles.linkButton} onClick={() => this.deleteSynSetAsk(synSet)}>
-                                                        <img src="../images/delete.svg" style={styles.dlImageSize} title="remove syn-set" alt="remove"/>
+                                                    <div className="link-button" onClick={() => this.deleteSynSetAsk(synSet)}>
+                                                        <img src="../images/delete.svg" className="image-size" title="remove syn-set" alt="remove"/>
                                                     </div>
-                                                </TableCell>
-                                            </TableRow>
+                                                </td>
+                                            </tr>
                                         )
                                     })
                                 }
-                                <TableRow>
-                                    <TableCell/>
-                                    <TableCell>
+                                <tr>
+                                    <td/>
+                                    <td>
                                         {this.isVisible() &&
-                                        <div style={styles.imageButton} onClick={() => this.newSynSet()}><img
-                                            style={styles.addImage} src="../images/add.svg" title="new syn-set"
+                                        <div className="image-button" onClick={() => this.newSynSet()}><img
+                                            className="image-size" src="../images/add.svg" title="new syn-set"
                                             alt="new syn-set"/></div>
                                         }
-                                    </TableCell>
-                                </TableRow>
+                                    </td>
+                                </tr>
 
-                            </TableBody>
+                            </tbody>
 
-                        </Table>
+                        </table>
 
-                        <TablePagination
+                        <Pagination
                             rowsPerPageOptions={[5, 10, 25]}
-                            style={theme === 'light' ? styles.tableLight : styles.tableDark}
+                            theme={theme}
                             component="div"
                             count={this.props.synset_total_size}
                             rowsPerPage={this.props.synset_page_size}
                             page={this.props.synset_page}
-                            backIconButtonProps={{
-                                'aria-label': 'Previous Page',
-                            }}
-                            nextIconButtonProps={{
-                                'aria-label': 'Next Page',
-                            }}
-                            onChangePage={(event, page) => this.props.setSynSetPage(page)}
-                            onChangeRowsPerPage={(event) => this.props.setSynSetPageSize(event.target.value)}
+                            onChangePage={(page) => this.props.setSynSetPage(page)}
+                            onChangeRowsPerPage={(rows) => this.props.setSynSetPageSize(rows)}
                         />
 
-                    </Paper>
+                    </div>
                 }
 
             </div>
